Revert "Clear filters (#76)"
This reverts commit aa1d585ceffc379fe1f71d405956fc481eae90e9.
