+++ src/components/HendelseTypeFilter.tsx
@@ -1,13 +1,9 @@
-import React, { ComponentPropsWithoutRef, useEffect } from 'react';
+import React, { useState, ComponentPropsWithoutRef } from 'react';
 import EkspanderbartPanel from 'nav-frontend-ekspanderbartpanel';
-import { useDispatch, useSelector } from 'react-redux';
 import { Checkbox } from 'nav-frontend-skjema';
-import countFilterAction from '../metrics/countFilterAction';
 import { PersonregisterState } from '../store/personregister/personregisterTypes';
 import { filtrerPersonregister } from '../utils/hendelseFilteringUtils';
-import { ApplicationState } from '../store';
-import { HendelseTypeFilters } from '../store/filters/filterReducer';
-import { updateHendelseFilterAction } from '../store/filters/filter_actions';
+import countFilterAction from '../metrics/countFilterAction';
 
 export const HendelseTekster: any = {
     MOTEBEHOV: 'Ønsker møte', // MØTEBEHOV - UBEHANDLET
@@ -19,6 +15,13 @@ export const HendelseTekster: any = {
 interface Props extends ComponentPropsWithoutRef<any> {
     onFilterChange: (filter: HendelseTypeFilters) => void;
     personRegister?: PersonregisterState;
+}
+
+export interface HendelseTypeFilters {
+    onskerMote: boolean;
+    svartMote: boolean;
+    ufordeltBruker: boolean;
+    ikkeIAktivitet: boolean;
 }
 
 const enkeltFilterFraTekst = (tekst: string, checked: boolean): HendelseTypeFilters => {
@@ -40,38 +43,31 @@ const lagNyttFilter = (forrigeFilter: HendelseTypeFilters, tekst: string, checke
     return filter;
 };
 
-const isCheckedInState = (state: HendelseTypeFilters, tekst: string): boolean => {
-    if (tekst === HendelseTekster.MOTEBEHOV) return state.onskerMote;
-    if (tekst === HendelseTekster.MOTEPLANLEGGER_SVAR) return state.svartMote;
-    if (tekst === HendelseTekster.UFORDELTE_BRUKERE) return state.ufordeltBruker;
-    if (tekst === HendelseTekster.IKKE_I_AKTIVITET) return state.ikkeIAktivitet;
-    return false;
-};
-
 interface CheckboksElement {
     tekst: string;
-    checked: boolean;
     key: string;
 }
 
-export default ({ className, personRegister }: Props) => {
+export default ({ onFilterChange: onValgteElementerChange, className, personRegister }: Props) => {
 
-    const dispatch = useDispatch();
-    const currentHendelseFilters = useSelector((state: ApplicationState) => state.filters.selectedHendelseType);
-
-    const updateHendelseFilterState = (filters: HendelseTypeFilters) => {
-        dispatch(updateHendelseFilterAction(filters));
+    const initialFilter: HendelseTypeFilters = {
+        onskerMote: false,
+        svartMote: false,
+        ufordeltBruker: false,
+        ikkeIAktivitet: false,
     };
+    const [filter, setFilter] = useState<HendelseTypeFilters>(initialFilter);
 
     const elementer = Object.keys(HendelseTekster).map((key) => {
         const tekst: string = HendelseTekster[key];
-        const checked = isCheckedInState(currentHendelseFilters, tekst);
-        return { key, tekst, checked } as CheckboksElement;
+        return { key, tekst } as CheckboksElement;
     });
 
     const onCheckedChange = (element: CheckboksElement, checked: boolean) => {
-        const nyttFilter = lagNyttFilter(currentHendelseFilters, element.tekst, checked);
-        updateHendelseFilterState(nyttFilter);
+        const nyttFilter = lagNyttFilter(filter, element.tekst, checked);
+        setFilter(nyttFilter);
+        onValgteElementerChange(nyttFilter);
+
         if (checked) {
             countFilterAction(element.tekst).next();
         }
@@ -98,7 +94,6 @@ const genererHendelseCheckbokser = (
         const labelNode = (<div>{checkboksElement.tekst} <strong>({antall})</strong></div>);
         return (<Checkbox
             label={labelNode}
-            checked={checkboksElement.checked}
             id={checkboksElement.key}
             key={checkboksElement.key}
             onChange={(e) => onCheckedChange(checkboksElement, e.target.checked)}
