+++ src/components/PersonFilter.tsx
@@ -1,12 +1,11 @@
 import React, { useState } from 'react';
 import Ekspanderbartpanel, { EkspanderbartpanelProps } from 'nav-frontend-ekspanderbartpanel';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { updateBirthDateFilter, updateCompaniesFilter } from '../store/filters/filter_actions';
 import BirthDateFilter from './filters/BirthDateFilter';
 import CompanyFilter from './filters/CompanyFilter';
 import styled from 'styled-components';
 import { PersonregisterState } from '../store/personregister/personregisterTypes';
-import { ApplicationState } from '../store';
 
 const texts = {
     panelTitle: 'Filter',
@@ -51,13 +50,11 @@ export default (props: PersonFilterProps) => {
         dispatch(updateCompaniesFilter(companies));
     };
 
-    const filters = useSelector((state: ApplicationState) => state.filters);
-
     return (
         <Ekspanderbartpanel apen={panelOpen} onClick={togglePanel} tittel={texts.panelTitle} >
             <SpacedFilters>
-                <CompanyFilter selectedOptions={filters.selectedCompanies} options={mapPersonregisterToCompanyList(personregister)} onSelect={onCompanyChange} />
-                <BirthDateFilter onSelect={onBirthDateChange} selectedDates={filters.selectedBirthDates} />
+                <CompanyFilter options={mapPersonregisterToCompanyList(personregister)} onSelect={onCompanyChange} />
+                <BirthDateFilter onSelect={onBirthDateChange} />
             </SpacedFilters>
         </Ekspanderbartpanel>
     );
