+++ src/components/toolbar/Toolbar.tsx
@@ -61,10 +61,7 @@ const InfoText = styled.div`
   flex-direction: row;
   font-size: 1em;
   font-weight: bold;
-  padding: 0.25em;
   padding-bottom: 0.5em;
-  box-sizing: border-box;
-  border: 2px solid transparent;
   >:not(:first-child) {
     margin-left: 0.5em;
   }
