+++ src/containers/EnhetensOversiktContainer.tsx
@@ -12,7 +12,7 @@ import AppSpinner from '../components/AppSpinner';
 import Sokeresultat from '../components/Sokeresultat';
 import { pushVeilederArbeidstakerForespurt } from '../store/veilederArbeidstaker/veilederArbeidstaker_actions';
 import { VeilederArbeidstaker } from '../store/veilederArbeidstaker/veilederArbeidstakerTypes';
-import SokeresultatFilter from '../components/HendelseTypeFilter';
+import SokeresultatFilter, { HendelseTypeFilters } from '../components/HendelseTypeFilter';
 import {
   Filterable,
   filtrerPersonregister,
@@ -29,7 +29,6 @@ import { AlertStripeWarning } from '../components/AlertStripeWarning';
 import { OverviewTabType } from '../konstanter';
 import { hentVeiledere } from '../store/veiledere/veiledere_actions';
 import PersonFilter from '../components/PersonFilter';
-import ClearFilters from '../components/filters/ClearFilters';
 
 const tekster = {
     feil: {
@@ -70,18 +69,13 @@ const HendelseFilterStyled = styled(SokeresultatFilter)`
   margin-bottom: 1rem;
 `;
 
-const ClearFiltersButton = styled(ClearFilters)`
-  display: flex;
-  justify-content: flex-end;
-  margin-bottom: .5em;
-`;
-
 interface Props {
   tabType?: OverviewTabType;
 }
 
 export default ({ tabType = OverviewTabType.ENHET_OVERVIEW  }: Props) => {
-
+  const initHendelseTypeFilter = {} as HendelseTypeFilters;
+  const [ hendelseTypeFilter, onHendelsesTypeChange ] = useState(initHendelseTypeFilter);
   const [ tekstFilter, onTekstFilterChange ] = useState('');
 
   const dispatch = useDispatch();
@@ -101,7 +95,6 @@ export default ({ tabType = OverviewTabType.ENHET_OVERVIEW  }: Props) => {
     selectedBirthDates,
     selectedVeilederIdents,
     selectedCompanies,
-    selectedHendelseTypeFilters,
   } = getPropsFromState(useSelector((state: ApplicationState) => state));
 
   useEffect(() => {
@@ -122,7 +115,7 @@ export default ({ tabType = OverviewTabType.ENHET_OVERVIEW  }: Props) => {
   const filteredEvents = new Filterable<PersonregisterState>({...allEvents.value})
     .applyFilter((v) => filterOnCompany(v, selectedCompanies))
     .applyFilter((v) => filterOnBirthDates(v, selectedBirthDates))
-    .applyFilter((v) => filtrerPersonregister(v, selectedHendelseTypeFilters))
+    .applyFilter((v) => filtrerPersonregister(v, hendelseTypeFilter))
     .applyFilter((v) => filtrerPaaFodselsnummerEllerNavn(v, tekstFilter));
 
   return (
@@ -132,11 +125,11 @@ export default ({ tabType = OverviewTabType.ENHET_OVERVIEW  }: Props) => {
       {noeErHentet && (
         <OversiktContainerInnhold>
           <SokeresultatFiltre>
-              <ClearFiltersButton />
               <TekstFilterStyled
                   onFilterChange={onTekstFilterChange}
               />
               <HendelseFilterStyled
+                  onFilterChange={onHendelsesTypeChange}
                   personRegister={allEvents.value}
                   tabType={tabType}
               />
@@ -169,5 +162,4 @@ const getPropsFromState = (state: ApplicationState) => ({
   selectedBirthDates: state.filters.selectedBirthDates,
   selectedVeilederIdents: state.filters.selectedVeilederIdents,
   selectedCompanies: state.filters.selectedCompanies,
-  selectedHendelseTypeFilters: state.filters.selectedHendelseType,
 });
