+++ src/store/filters/filter_actions.ts
@@ -1,11 +1,7 @@
-import { HendelseTypeFilters } from './filterReducer';
-
 export enum FilterActionTypes {
     UPDATE_BIRTH_DATE = 'UPDATE_BIRTH_DATE',
     UPDATE_VEILDERER_IDENTS = 'UPDATE_VEILEDER_IDENTS',
     UPDATE_COMPANIES = 'UPDATE_COMPANIES',
-    UPDATE_HENDELSE_FILTER = 'UPDATE_HENDELSE_FILTER',
-    RESET_ALL_FILTERS = 'RESET_ALL_FILTERS',
 }
 
 interface UpdateBirthDateFilter {
@@ -18,31 +14,13 @@ interface UpdateVeilederIdentsFilter {
     selectedVeilederIdents: string[];
 }
 
+export type FilterAction =
+    UpdateBirthDateFilter | UpdateVeilederIdentsFilter | UpdateCompaniesFilter;
+
 interface UpdateCompaniesFilter {
     type: FilterActionTypes.UPDATE_COMPANIES;
     selectedCompanies: string[];
 }
-
-interface UpdateHendelseFilter {
-    type: FilterActionTypes.UPDATE_HENDELSE_FILTER;
-    filter: HendelseTypeFilters;
-}
-interface ResetAllFilters {
-    type: FilterActionTypes.RESET_ALL_FILTERS;
-}
-
-export type FilterAction =
-    UpdateBirthDateFilter
-    | UpdateVeilederIdentsFilter
-    | UpdateCompaniesFilter
-    | UpdateHendelseFilter
-    | ResetAllFilters
-    ;
-
-export const updateHendelseFilterAction = (hendelseFilter: HendelseTypeFilters) => ({
-    filter: hendelseFilter,
-    type: FilterActionTypes.UPDATE_HENDELSE_FILTER,
-}) as UpdateHendelseFilter;
 
 export const updateBirthDateFilter = (birthDates: string[]) => ({
     type: FilterActionTypes.UPDATE_BIRTH_DATE,
@@ -53,12 +31,7 @@ export const updateVeilederIdentsFilter = (veilederIdents: string[]) => ({
     type: FilterActionTypes.UPDATE_VEILDERER_IDENTS,
     selectedVeilederIdents: veilederIdents,
 }) as UpdateVeilederIdentsFilter;
-
 export const updateCompaniesFilter = (companies: string[]) => ({
     type: FilterActionTypes.UPDATE_COMPANIES,
     selectedCompanies: companies,
 }) as UpdateCompaniesFilter;
-
-export const resetAllFilters = () => ({
-    type: FilterActionTypes.RESET_ALL_FILTERS,
-}) as ResetAllFilters;
